Found query length
Ok I got the query length and the function now sorts by query coverage.
What I did basically was use the gene id and used entrez (we used it
before to get protein id) to get the protein sequence which I took the
length of to get the protein sequence length of the original.