Added parser for fasta files
all the original code for the file has been commented out underneath
the code I added. The result of the code places the information about
each fasta into one string that is in a list. Index the list to
retrieve the strings, a counter has also been used to tell you how big
the list is